==> Wyoming/WaterSupply_AggregatedArea/0_WYAggregatedDataPreprocess.ipynb <==
# Working with WY Aggregated Data

Preprocessing input data for a smoother upload experience of the state data to the WaDE 2.0 database.
Using geopandas to read in shp file, and coverting to WKT for ReportingUnit geometry.

#Needed Libararies

# working with data
import os
import numpy as np
import pandas as pd
import geopandas as gpd

# visulizaiton
import matplotlib.pyplot as plt
import seaborn as sns

# Cleanup
from datetime import datetime
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

#Setting work directory, reading inputs, creating dataframe
workingDir = "G:/Shared drives/WaDE Data/Wyoming/AggregatedAmounts/RawInputData"
os.chdir(workingDir)

# Water Use Data

# Input Dataframe
fileInput = "WYAggData_input.csv"
df = pd.read_csv(fileInput)
print(len(df))
df.head(1)

# temp database #1: base elements
droplist = ["2000", "2001", "2002", "2003", "2004", "2005", "2006",
            "2007", "2008", "2009", "2010", "2011", "2012", "2013",
            "2014", "2015", "2016", "2017", "2018"]

dftempK = df.copy()
dftempK = dftempK.drop(columns=droplist).reset_index()
print(len(dftempK))
dftempK.head(1)

# temp database #2: time series amounts
# restructure

dftempY = df.copy()
dftempY = dftempY.drop(columns=['Water Use by Basin', 'UseType', 'Source'])

dftempY = dftempY.T.unstack().reset_index(level=1, name='Amount').rename(columns={'level_1':'ReportYear'})[['Amount','ReportYear']]

print(len(dftempY))
dftempY.head(1)

# Combine database #1 to database #2.
dftempY['Water Use by Basin'] = dftempK['Water Use by Basin']
dftempY['UseType'] = dftempK['UseType']
dftempY['Source'] = dftempK['Source']

print(len(dftempY))
dftempY.head(1)

# Fixing UseType errors
# Uncessary spaces & white space.

def fixUseType(val):
    val = str(val).strip() 
    if val == "" or pd.isnull(val):
        val = "Unspecified"
    else:
        # Cleaning text / simple search format
        val = val.strip()
        val = val.replace("  ", " ")
        val = val.replace(",", "")
        val = val.replace(".", "")
        val = val.replace(";", "")
        val = val.replace("-", "")
        val = val.replace("/", "")
        val = val.replace("(", "")
        val = val.replace(")", "")

        return val

dftempY['UseType'] = dftempY.apply(lambda row: fixUseType(row['UseType']), axis=1)
dftempY['UseType'].unique()

# Timeframe start & end
dftempY['in_TimeframeStart'] = dftempY['ReportYear'].astype(str) +  "/01/01" 
dftempY['in_TimeframeEnd'] = dftempY['ReportYear'].astype(str) + "/12/31" #cheat for now and use value of 28 for day


# Convert History Year to YYYY-MM-DD format.
dftempY['in_TimeframeEnd'] = pd.to_datetime(dftempY['in_TimeframeEnd'], errors = 'coerce')
dftempY['in_TimeframeEnd'] = pd.to_datetime(dftempY["in_TimeframeEnd"].dt.strftime('%m/%d/%Y'))

dftempY['in_TimeframeStart'] = pd.to_datetime(dftempY['in_TimeframeStart'], errors = 'coerce')
dftempY['in_TimeframeStart'] = pd.to_datetime(dftempY["in_TimeframeStart"].dt.strftime('%m/%d/%Y'))

dftempY.head(1)

## WaDE Custom Elements (due to missing sate info)

# temp watersource type

def tempWaterSourceType(WST):
    WST = str(WST).strip()
    if WST == "Ground Water":
        outString = "Groundwater"
    else:
        outString = WST
   
    return outString

dftempY['tempWST'] = dftempY.apply(lambda row: tempWaterSourceType(row['Source']), axis=1)
dftempY['tempWST'].unique()

# Creating VariableSpecificCV

def createVariableSpecificCV(UT, WST):
    UT = str(UT).strip()
    WST = str(WST).strip()
    
    outString = "Consumptive Use_Annual_" +  UT + "_" + WST
    
    return outString

dftempY['in_VariableSpecificCV'] = dftempY.apply(lambda row: createVariableSpecificCV(row['UseType'], row['tempWST']), axis=1)
dftempY['in_VariableSpecificCV'].unique()

# Creating WaDE Custom reporting u nit native ID for easy identificaiion
# ----------------------------------------------------------------------------------------------------

# Create temp ReportingUnit native id dataframe of unique site.
def assignReportNID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDEWY_RU" + string1
    return outstring

dfReportNativeID = pd.DataFrame()

dfReportNativeID['in_ReportingUnitName'] = dftempY['Water Use by Basin']
dfReportNativeID = dfReportNativeID.drop_duplicates()

dftemp = pd.DataFrame(index = dfReportNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfReportNativeID['in_ReportingUnitNativeID'] = dftemp.apply(lambda row: assignReportNID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom ReportingUnit native id
def retrieveReportNID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfReportNativeID.loc[(dfReportNativeID['in_ReportingUnitName'] == A), 'in_ReportingUnitNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

dftempY['in_ReportingUnitNativeID'] = dftempY.apply(lambda row: retrieveReportNID(row['Water Use by Basin']), axis=1)
dftempY['in_ReportingUnitNativeID'].unique()

# Creating WaDE Custom water source native ID for easy water source identification
# ----------------------------------------------------------------------------------------------------

# Create temp WaterSourceNativeID dataframe of unique water source.
def assignWaterSourceNativeID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDEWY_WS" + string1
    return outstring

dfWaterSourceNativeID = pd.DataFrame()
dfWaterSourceNativeID['in_WaterSourceTypeCV'] = dftempY['Source']
dfWaterSourceNativeID = dfWaterSourceNativeID.drop_duplicates()

dftemp = pd.DataFrame(index=dfWaterSourceNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfWaterSourceNativeID['in_WaterSourceNativeID'] = dftemp.apply(lambda row: assignWaterSourceNativeID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom water source native ID
def retrieveWaterSourceNativeID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfWaterSourceNativeID.loc[(dfWaterSourceNativeID['in_WaterSourceTypeCV'] == A), 'in_WaterSourceNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

dftempY['in_WaterSourceNativeID'] = dftempY.apply(lambda row: retrieveWaterSourceNativeID(row['Source']), axis=1)
dftempY['in_WaterSourceNativeID'].unique()

# Shapefile Data

# Shapefile input
shpInput = "WYBasinShapefile/BasinShapefile.shp"
BasinShape = gpd.read_file(shpInput)
print(len(BasinShape))
BasinShape.head(1)

#Create geo output
columnsList = ['RU_Name', 'geometry']
dfshape = pd.DataFrame(columns=columnsList)
dfshape['RU_Name'] = BasinShape['Reportin_3']
dfshape['geometry'] = BasinShape['geometry']
dfshape = dfshape.drop_duplicates(subset=None, keep='first', inplace=False, ignore_index=False)
print(len(dfshape))
dfshape.head(1)

### Inspect Output Data & Export

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dftempY.dtypes)

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dfshape.dtypes)

# Export out to CSV.
dftempY.to_csv('P_wyAggMaster.csv', index=False) # The output.
dfshape.to_csv('P_wyGeometry.csv', index=False) # The output geometry.
